+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Hike ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added the ability to deduplicate the history.
+
 ## v0.12.0
 
 **Released: 2025-04-30**